Added Preferences, fixed naming, Biaru's patch!
Added a panel to the user preferences to give users easy access to the time threshold - how long do you hold before navigation / how short is a click for the context menu, and a second panel for Biaru's distance threshold - how far you move the cursor to trigger navigation.
Also fixed up inconsistencies in naming: the addon used to be called UnrealRightClick, but I figured anyone who didn't use UE might not know what it does, so I changed the repo to RightMouseNavigation, and then made sure that naming convention was kept throughout the code itself.
Biaru's pull request fixed a bug that had been bugging me - the context menu and cursor always snapped to the centre of the screen after navigation. With the fix, the cursor position is stored, and RESTORED after navigation, which has the knock on effect of making the context menus show up where they're supposed to!
Huge thanks to Biaru for making this addon better!