Resolved "Disable/Reenable bug" on opening a blend
Reported by Ares, this one was a bit of a bear, mostly because I was looking in the wrong spot.

I tried importing time to set a delay before registration, tried using threading to monitor the keymap registration, tried enabling "Autorun Python Scripts", and nothing worked. The addon worked when Blender was opened first, but not when a .blend file was clicked before the program was open.

Solved by iceythe Kaio on Blender Artists:

https://blenderartists.org/t/why-wont-this-addon-retain-keymap-user-preferences/1164306

When you store keymaps in addon keymaps they are dumped between blender sessions, but when you store them in user keymaps, they remain between sessions, and, thus, don't need to be manually disabled and re-enabled every time you double click a blend file.