Cross Platform Support, Big Refactor, Node Editors
I removed as much of the Ctypes code that relied upon the windll as I could, replacing it with Blender native functions.
The only place it couldn't be removed was from the fake Right Mouse Release, which is only required on Windows anyway, so it worked itself out. (Still not sure why Windows needs it but Linux does not).
Haven't tested on Mac yet, but theoretically it should work. (famous last words)

Moved all the menu logic into a single dict that maps the context mode to the menu that the mode should call.
This means that there is no string formatting happening when attempting to call a menu, which gave inconsistent performance and added needless overhead.

You can now enable it for use in Node Editors - you can pan the view and open the Add/Search menu with RIght Mouse.

Also cleaned up the Preferences panel, as well as renaming variables, removing whitespace and other small changes.