Keymap, Cursor Reset Fix, Distance Removed
Fixed the keymap registration (again, and hopefully for the last time), I had been trying to make all the keymap changes on a single map, not realizing that I should have been switching between the addon keymap that I had created and the user's active keymap.

Cursor Reset behavior seems to have been updated in Blender so that the cursor will remember its last screen position. This is the 'patched' behavior that this addon previously had implemented, which made the Cursor Reset setting redundant.
I have implemented the previous 'cursor centering' behavior as the new 'Cursor Reset' behavior.
The settings have flipped, but now the old behavior remains.

The distance threshold did not have any effect on the addon, so it has been removed. If this is something that is missed, I can see about reimplementing it.