Blender 4.2 compatibility update
Removed `bl_info`, replaced with `blender_manifest.toml`
Updated Node Editor mode so that when a node is selected you will open the default context menu, but when no nodes are selected you will open the node search. I feel this is a practicality update, and the alternative would be an additional keybind, like `Ctrl + Right Mouse` to get the standard context menu.
If you have a strong opinion about this, create an issue and solutions can be discussed.
Thanks!